backend/requirements.txt
azure-functions
requests
sendgrid
python-dotenv
azure-storage-blob
mailersend




